Modify --md to take a minimum dimension in pixels, rather than a maximum decomposition degree.
darcs-hash:20080118195920-d1b56-3f1e8a8c372d88770af9f9824febe181ef481863.gz

+++ d2/align.h
@@ -3401,9 +3401,9 @@ public:
 	}
 
 	/*
-	 * Multi-alignment subdivision limit
+	 * Multi-alignment dimension minimum
 	 */
-	static void ma_decomp(unsigned int value) {
+	static void ma_decomp(double value) {
 		assert(0);
 	}
